requirements.txt
# GraphyloVar Requirements
# Auto-generated by rebuttal/generate_requirements.py
# Scanned: /home/mcb/users/dlim63/GraphyloVar
#

# Core ML
tensorflow==2.4.0
numpy==1.20.0
pandas==1.3.4
scikit-learn==1.3.2

# Graph Neural Networks
spektral==1.2.0
networkx==3.1

# Bioinformatics
pysam==0.22.1

# Visualization
matplotlib==3.3.4

# Utilities
PyYAML==6.0
tqdm==4.67.3
requests==2.32.4

# Other
setuptools==58.0.4
